Time: 0 ms (100.00%), Space: 6.6 MB (37.93%) - LeetHub

+++ 2000-reverse-prefix-of-word/2000-reverse-prefix-of-word.cpp
@@ -4,7 +4,7 @@ public:
         
         if(word.find(ch) != string::npos)
         {
-            size_t n = word.find(ch);
+            int n = word.find(ch);
             reverse(word.begin(), word.begin()+n+1);
         }
